+++ NoPhysicsLibrary/Source/PhysRect.h
@@ -6,11 +6,11 @@ struct PhysRect
 {
 	PhysRect() {}
 	PhysRect(float x, float y, float w, float h) : x(x), y(y), w(w), h(h) {}
-	PhysRect(int   x, int   y, int   w, int   h) : x(x), y(y), w(w), h(h) {}
+	PhysRect(int   x, int   y, int   w, int   h) : x((float)x), y((float)y), w((float)w), h((float)h) {}
 	PhysRect(float x, float y, PhysVec size) : x(x), y(y), w(size.x), h(size.y) {}
-	PhysRect(int   x, int   y, PhysVec size) : x(x), y(y), w(size.x), h(size.y) {}
+	PhysRect(int   x, int   y, PhysVec size) : x((float)x), y((float)y), w(size.x), h(size.y) {}
 	PhysRect(PhysVec position, float w, float h) : x(position.x), y(position.y), w(w), h(h) {}
-	PhysRect(PhysVec position, int   w, int   h) : x(position.x), y(position.y), w(w), h(h) {}
+	PhysRect(PhysVec position, int   w, int   h) : x(position.x), y(position.y), w((float)w), h((float)h) {}
 	PhysRect(PhysVec position, PhysVec size) : x(position.x), y(position.y), w(size.x), h(size.y) {}
 
 	void operator*=(float value);
